web-app/ml-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
torch==2.9.1
torchvision==0.24.1
Pillow==11.0.0

